fix must_be_fixed compilation warnings in seqregion.h/cpp; modelbase.h

+++ alignment/seqregion.h
@@ -88,7 +88,7 @@ class SeqRegion : public Mutation {
    */
   cmaple::RealNumType getLH(int pos) const {
       assert(type == TYPE_O);
-      return (*likelihood)[pos];
+      return (*likelihood)[(unsigned long) pos];
   }
 
   /**
